fixed firebase rules for proptypes
