+++ pages/index.js
@@ -1,25 +1,40 @@
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useEffect, useState } from 'react';
+import Link from 'next/link';
 import { Button } from 'react-bootstrap';
-import { signOut } from '../utils/auth';
+import { getMembers } from '../api/memberData';
 import { useAuth } from '../utils/context/authContext';
+import MembCard from '../components/MembCard';
 
 function Home() {
+  // TODO: Set a state for books
+  const [members, setMembers] = useState([]);
+
+  // TODO: Get user ID using useAuth Hook
   const { user } = useAuth();
 
+  // TODO: create a function that makes the API call to get all the books
+  const getAllTheMembers = () => {
+    getMembers(user.uid).then(setMembers);
+  };
+
+  // TODO: make the call to the API to get all the books on component render
+  useEffect(() => {
+    getAllTheMembers();
+  }, []);
+
   return (
-    <div
-      className="text-center d-flex flex-column justify-content-center align-content-center"
-      style={{
-        height: '90vh',
-        padding: '30px',
-        maxWidth: '400px',
-        margin: '0 auto',
-      }}
-    >
-      <h1>Hello {user.displayName}! </h1>
-      <p>Click the button below to logout!</p>
-      <Button variant="danger" type="button" size="lg" className="copy-btn" onClick={signOut}>
-        Sign Out
-      </Button>
+    <div className="text-center my-4">
+      <Link href="/member/new" passHref>
+        <Button>Add</Button>
+      </Link>
+      <div className="d-flex flex-wrap">
+        {/* TODO: map over members here using MembCard component */}
+        {members.map((member) => (
+          <MembCard key={member.firebaseKey} memberObj={member} onUpdate={getAllTheMembers} />
+        ))}
+      </div>
+
     </div>
   );
 }
