+++ pages/member/edit/{firebaseKey}.js
@@ -1,0 +1,19 @@
+import React, { useEffect, useState } from 'react';
+import { useRouter } from 'next/router';
+import { getSingleMember } from '../../../api/memberData';
+import MemberForm from '../../../components/forms/memberForm';
+
+export default function EditMember() {
+  const [editItem, setEditItem] = useState({});
+  const router = useRouter();
+  // TODO: grab the firebasekey
+  const { firebaseKey } = router.query;
+
+  // TODO: make a call to the API to get the book data
+  useEffect(() => {
+    getSingleMember(firebaseKey).then(setEditItem);
+  }, [firebaseKey]);
+
+  // TODO: pass object to form
+  return (<MemberForm obj={editItem} />);
+}
