--- pages/members.js
@@ -1,42 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import Link from 'next/link';
-import { Button } from 'react-bootstrap';
-import { getMembers } from '../api/memberData';
-import { useAuth } from '../utils/context/authContext';
-import MembCard from '../components/MembCard';
-
-function CatMembers() {
-  // TODO: Set a state for books
-  const [members, setMembers] = useState([]);
-
-  // TODO: Get user ID using useAuth Hook
-  const { user } = useAuth();
-
-  // TODO: create a function that makes the API call to get all the books
-  const getAllTheMembers = () => {
-    getMembers(user.uid).then(setMembers);
-  };
-
-  // TODO: make the call to the API to get all the books on component render
-  useEffect(() => {
-    getAllTheMembers();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return (
-    <div className="text-center my-4">
-      <Link href="/member/new" passHref>
-        <Button>Add New Member</Button>
-      </Link>
-      <div className="d-flex flex-wrap">
-        {/* TODO: map over books here using BookCard component */}
-        {members.map((member) => (
-          <MembCard key={member.firebaseKey} memberObj={member} onUpdate={getAllTheMembers} />
-        ))}
-      </div>
-
-    </div>
-  );
-}
-
-export default CatMembers;
