+++ utils/data/members.json
@@ -1,38 +1,38 @@
 {
   "-Not19Um9StBf5ywl3FT" : {
-    "firebaseKey": "",
-    "image": "null",
-    "name": "null",
-    "role": "null",
-    "uid": ""
+    "firebaseKey": "-Not19Um9StBf5ywl3FT",
+    "image": "https://i.imgur.com/BwXXWUd.jpg",
+    "name": "Stalin",
+    "role": "Dictator",
+    "uid": "2Zbq7O3PbYg83fcnc8op3IL8nRz1"
     },
   "-Not19UnGxgZgRwCqnzm" : { 
-    "firebaseKey": "",
-    "image": "null",
-    "name": "null",
-    "role": "null",
-    "uid": ""
+    "firebaseKey": "-Not19UnGxgZgRwCqnzm",
+    "image": "https://i.imgur.com/HdbUkHQ.jpg",
+    "name": "Margaret",
+    "role": "Bookseller",
+    "uid": "2Zbq7O3PbYg83fcnc8op3IL8nRz1"
     },
   "-Not19UpsI9KeCJJUUgO" : { 
-    "firebaseKey": "",
-    "image": "null",
-    "name": "null",
-    "role": "null",
-    "uid": ""
+    "firebaseKey": "-Not19UpsI9KeCJJUUgO",
+    "image": "blob:https://imgur.com/7a531c2c-1c1e-4a23-a21e-86d66a87e706",
+    "name": "Chrysanthemum",
+    "role": "Brooklyn Hipster",
+    "uid": "2Zbq7O3PbYg83fcnc8op3IL8nRz1"
     },
   "-Not19Uqvwx-xWe38lZl" : { 
-    "firebaseKey": "",
-    "image": "null",
-    "name": "null",
-    "role": "null",
-    "uid": ""
+    "firebaseKey": "-Not19Uqvwx-xWe38lZl",
+    "image": "blob:https://imgur.com/ed231649-c1d4-4475-b2ae-c9e31d312f00",
+    "name": "Bertha",
+    "role": "Medicare Member",
+    "uid": "2Zbq7O3PbYg83fcnc8op3IL8nRz1"
     },
   "-Not19UrL2JfJaNeTXwV" : { 
-    "firebaseKey": "",
-    "image": "null",
-    "name": "null",
-    "role": "null",
-    "uid": ""
+    "firebaseKey": "-Not19UrL2JfJaNeTXwV",
+    "image": "https://i.imgur.com/NV8hXQz.jpg",
+    "name": "Hacketha",
+    "role": "PC Hacker",
+    "uid": "2Zbq7O3PbYg83fcnc8op3IL8nRz1"
     }
   }
